requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy==2.0.20
alembic>=1.12.1
asyncpg==0.28.0
aiokafka>=0.8.1
python-dotenv>=1.0.0
redis>=5.0.1
prometheus-client>=0.17.1
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx==0.23.3
pre-commit>=3.3.3
black>=23.7.0
isort>=5.12.0
